split field mod
